+++ tests/database/migrations/2020_07_09_000000_create_testbench_activities_table.php
@@ -1,0 +1,26 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class CreateTestbenchActivitiesTable extends Migration
+{
+    public function up(): void
+    {
+        Schema::create('activities', function (Blueprint $table): void {
+            $table->increments('id');
+            $table->bigInteger('user_id');
+            $table->morphs('content');
+            $table->timestamps();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     */
+    public function down(): void
+    {
+        Schema::drop('activities');
+    }
+}