+++ CHANGELOG.md
@@ -13,6 +13,11 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 - Add artisan command `lighthouse:cache` to compile GraphQL AST https://github.com/nuwave/lighthouse/pull/1451
 
+### Added
+
+- Add middleware `\Nuwave\Lighthouse\Support\Http\Middleware\LogGraphQLQueries` that logs every incoming
+  GraphQL query https://github.com/nuwave/lighthouse/pull/1454
+
 ## 4.15.0
 
 ### Added