+++ CHANGELOG.md
@@ -15,6 +15,11 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Add middleware `\Nuwave\Lighthouse\Support\Http\Middleware\LogGraphQLQueries` that logs every incoming
   GraphQL query https://github.com/nuwave/lighthouse/pull/1454
 
+### Changed
+
+- Publish config file with tag `lighthouse-config` and default schema with tag `lighthouse-schema`
+  instead of the previously used tags `config` and `schema` https://github.com/nuwave/lighthouse/issues/1489
+
 ### Fixed
 
 - Ensure the `@count` directive works properly with polymorphic relations https://github.com/nuwave/lighthouse/pull/1466